Create audio_silence.py
This code opens an audio stream using PyAudio, reads chunks of audio data, and checks if any of the data contains zeros. If zeros are detected, it prints a message indicating that a dropout has been detected. Note that this code is just a starting point and will need to be customized to suit your specific needs. You will also need to integrate it with the streaming platform that your radio station uses to receive and analyze the stream.